server/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
binaryornot==0.4.4
certifi==2024.6.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.40.0
cookiecutter==2.6.0
decorator==5.1.1
Django==4.2
django-allauth==0.63.3
django-browser-reload==1.12.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==24.2
django-tailwind==3.6.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
executing==2.0.1
frozenlist==1.4.1
genai==2.1.0
idna==3.7
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.0.5
openai==0.27.10
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
tabulate==0.9.0
text-unidecode==1.3
tiktoken==0.3.3
tqdm==4.66.4
traitlets==5.14.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
urllib3==2.2.1
wcwidth==0.2.13
yarl==1.9.4
